style: Separate utils files
